Initial project import
Maven multi-module project with a setup involving multiple AspectJ ITD
aspects, each in a separate module, which are applied to the same
application module. The main requirement for me setting this up was to
avoid a dependency chain between the aspect modules, i.e. each aspect
module should depend on the core module, but not know anything about any
other aspect modules.

Parent is 'aspectj-multi-module', child modules are:
  - data-objects (application class, no aspects)
    no dependencies
  - aspect-javabean (Java Bean-related ITD aspect)
    depends on data-objects
  - aspect-json (JSON-related ITD aspect)
    depends on data-objects
  - data-objects-aspectj (weave aspects into core classes)
    depends on data-objects, aspect-javabean, aspect-json
  - client-jar (create JAR of JARs containing core + aspect code, sample
    class with main method using aspect-enhanced core class and AspectJ
    runtime)
    depends on data-objects-aspectj